+++ packages/crayon/src/plugin/generators/__tests__/padding.test.ts
@@ -1,6 +1,6 @@
 import { paddingBottom, paddingLeft, paddingRight, paddingTop } from '../padding'
 
-jest.mock('../../tailwindUtils')
+jest.mock('../../../utils/tailwindUtils')
 
 describe.each(
   [
