+++ packages/crayon/src/plugin/plugin.ts
@@ -1,7 +1,7 @@
 import { TailwindUtils, UserConfig } from '../types'
 import resolveThemes from '../utils/resolveThemes'
 import { processThemes } from './processThemes'
-import { storeTailwindUtils } from './tailwindUtils'
+import { storeTailwindUtils } from '../utils/tailwindUtils'
 
 export default function plugin (config: UserConfig) {
   return function (tailwindUtils: TailwindUtils): void {
