+++ packages/crayon/src/utils/__tests__/generateComponentName.test.ts
